requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==3.1.0
Markdown==3.3.3
Pillow==8.0.1
pip==20.3
psycopg2-binary==2.8.6
pytz==2020.4
setuptools==41.2.0
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.4.0
